Getting rid of offsets!! Alter query code to cast ISO and UNIX dates to native datetimes and do timezone conversion in the database. Much faster and more accurate.
This will mean this version of the code will not work for databases that are not able to do native timezone conversion, but even inexpensive shared hosts will probably have the MYSQL database timezone tables installed, and PostgreSQL has good timezone handling enabled by default.

The MYSQL timezone conversion code should be accurate, the PostgreSQL code looks right from the documentation but needs to be verified.

The offset fields can now be dropped from the Date fields since we only need the timezone name to do the conversion.

I probably missed some places that need to be adjusted for this change, but this should get most of the necessary fixes in place.