More work to get the generic date argument handler pulled completely out of the Calendar module and into the Date API in a way that it can be used for any Views date field in any view.
Rework the calendar handling to use the Date API argument in calendar displays and  navigation. Keep the calendar displays out of the UI and instead create a default calendar view that has everything set up correctly to reduce confusion and clutter in the UI and to make it easier to create a calendar.

More work is still needed to get this finished and working right.